Rename service and introduce install scripts
It was needed to rename the service, so it fits to the original udev script.
The intention is to make things smarter, which the official installer isn't doing well, and nothing more.

Later I'll care about using debconf to show the license and let the user agree on it or not.
Depending on the outcome the installation will be aborted.

Signed-off-by: Thomas Karl Pietrowski <thopiekar@gmail.com> (github: thopiekar)